Base medication list screen

+++ medicationapp/lib/pages/medication/medication_data.dart
@@ -1,7 +1,7 @@
 class MedicationType {
   String name;
   int quantityRemaining;
-  int dosage;
+  double dosage;
 
   MedicationType(this.name, this.quantityRemaining, this.dosage);
 }
